Move test to proper location
